[Recipe] Distributed work queue for long-running tasks (#1)
### Motivation

A distributed work queue for long-running tasks. A particular challenge
with long-running tasks is that such a system is
more susceptible to consumer connection interruptions. These
interruptions would result in message redeliveries and
duplicate processing of the tasks. This system seeks to avoid this
problem by maintaining a global task state in a
separate topic.

The work scheduler supports task retries, detection of non-progressing
work, and the cancellation of work that has
exceeded a time-based processing budget.

### Modifications

* Implementation of a distributed long-running task work queue.
* Unit test suite
* Integration test
* Documentation (README)

### Verifying this change

- [x] Make sure that the change passes the CI checks.

Co-authored-by: teabot <elliot.west@streamnative.io>
Co-authored-by: Elliot West <teabot@gmail.com>
Co-authored-by: Andras Beni <andras.beni@streamnative.io>